refactor: reorganize enums and types for better structure and clarity
