+++ app/src/lib/exercise_data/calisthenics.ts
@@ -1,5 +1,6 @@
+import { Muscles } from '$lib/muscles';
 import type { ExerciseDetails } from '../types';
-import { Muscles } from '../enums';
+
 
 /**
  * A collection of calisthenics exercises that can be used to generate workouts.
