+++ app/src/lib/exercise_data/dumbbells.ts
@@ -1,5 +1,6 @@
+import { Equipment } from '$lib/equipment';
+import { Muscles } from '$lib/muscles';
 import type { ExerciseDetails } from '../types';
-import { Muscles, Equipment } from '../enums';
 
 /**
  * A collection of dumbbell exercises that can be used to generate workouts.
