+++ app/src/lib/exercise_data/pilates.ts
@@ -1,5 +1,5 @@
+import { Muscles } from '$lib/muscles';
 import type { ExerciseDetails } from '../types';
-import { Muscles } from '../enums';
 
 /**
  * A collection of common Pilates exercises that can be used to generate workouts.
