+++ app/src/lib/musclesData.ts
@@ -1,0 +1,3 @@
+import type { MusclesDetails } from './types';
+import { Muscles } from './joints';
+
